feat: remove TrustbridgeContract and related files, add test suite for escrow functionality
