--- trustbridge_contract/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "trustbridge_contract"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-ink = { version = "5.1.1", default-features = false }
-scale = { package = "parity-scale-codec", version = "3", default-features = false, features = ["derive"] }
-scale-info = { version = "2.0", default-features = false }
-
-[features]
-std = [
-    "ink/std",
-    "scale/std",
-    "scale-info/std",
-]
-[lib]
-path = "../trustbridge_contract/trustbridge_escrow.rs"
-
-[profile.release]
-lto = "fat"
-opt-level = "z"
